requirements.txt
# ----------------------
# Flask
# ----------------------
astroid==2.4.0
click==7.1.2
dnspython==1.16.0
email-validator==1.1.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pkg-resources==0.0.0
pylint==2.5.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1

# ----------------------
# Production Server
# ----------------------
gunicorn
